Use a custom directory for session data
This gives us more flexibility w.r.t. permissions (we can make sure
it's owned by the same user / group as Apache).

This change is needed become as of the upgrade of the ugprade to the
Apache Cookbook 2.0.0, we changed the group (it used to be www-data /
apache, now it's scalr) that the Apache process runs as.