Enable the SSH Launcher on Scalr 5.x
Mindterm needs to be disabled by default on Scalr 4.x (because it
requires a license), but the SSH Launcher can be enabled on Scalr 5.x
(because it's ours).

The configuration key is the same on both versions
(ui.mindterm_enabled), and defaults to false, so we just need to
override it in Scalr 5.x.