Add a match-version command
This lets us ensure that we recreate configuration files in the
dist/install.sh script after an update.

As a general rule dist/install.sh is for first-time installs, so it's
likely that if someone is updating and re-running it, they're doing so
because something is broken (which might be the configuration file!).